move everything from predict_client submodule into redis_consumer
