+++ pytest.ini
@@ -11,7 +11,7 @@ filterwarnings =
 # Do not run tests in the build folder
 norecursedirs= 
     build
-    redis_consumer/predict_client/pbs
+    redis_consumer/pbs
 
 
 # PEP-8 The following are ignored:
