+++ redis_consumer/__init__.py
@@ -27,12 +27,13 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-from redis_consumer import predict_client
+from redis_consumer import grpc_client
 from redis_consumer import processing
 from redis_consumer import consumers
 from redis_consumer import settings
 from redis_consumer import storage
 from redis_consumer import utils
+from redis_consumer import pbs
 
 del absolute_import
 del division
