+++ redis_consumer/pbs/predict_pb2_grpc.py
