+++ redis_consumer/pbs/resource_handle_pb2_grpc.py
