+++ redis_consumer/pbs/tensor_shape_pb2_grpc.py
